- added new tests for EnvironmentServiceImplTest and AnimalEnvironmentServiceImplTest - added IllegalArgumentException throwing for null arguments in AnimalEnvironmentService and EnvironmentService - updated equals() for Animal and AnimalEaten entity

+++ DominatingSpeciesService/src/test/java/cz/muni/fi/pa165/dominatingspecies/facade/EnvironmentFacadeImplTest.java
@@ -226,7 +226,7 @@ public class EnvironmentFacadeImplTest {
     }
     
     private void assertAnimalDataEquals(AnimalBriefDTO expected, AnimalDetailDTO actual) {
-        assertEquals(expected.getId(), actual.getId());
+        assertEquals((Long)expected.getId(), (Long)actual.getId());
         assertEquals(expected.getName(), actual.getName());
         assertEquals(expected.getSpecies(), actual.getSpecies());
     }
